Role creation
This first version does a very similar sequence of operations to
tripleo-quickstart undercloud installation.
Omitting all the parts related to the virtual, treats virthost and
undercloud as the same machine, uses a conditional on variable release
for valuing STABLE_RELEASE and does these tasks:

- Provide baremetal machine which will become both virthost/undercloud;
- Prepare the undercloud copying all necessary files;
- Execute the pre installation script;
- Install the undercloud;
- Take overcloud images;

The baremetal-undercloud.yml inside the playbooks directory gives a way
to invoke this role.

For this role to work with all the others to get an environment ready to
use these baremetal patches must be applied to the respective roles:

- ansible-role-tripleo-inventory [1]
- ansible-role-tripleo-undercloud-post [2]
- ansible-role-tripleo-overcloud [3]

[1] https://review.gerrithub.io/#/c/274491/
[2] https://review.gerrithub.io/#/c/275333/
[3] https://review.gerrithub.io/#/c/275011/

Change-Id: I773d0d2f36284c41c646410f47c357323262efc0