Extra logs as text files, add yum package list
Move the package list outputs to /var/log/extra like the rest of the
addition outputs and change the extension of all of them to a standard
txt to help opening them in text editors.

Also add yum list installed output to help figure out the source repo
of all installed packages, not just their version.

Change-Id: I99ca3ce07f1fbba126cd06a1bc84099f3825d81f