Delete images and networks on exit
If you want to run more than one test the actual way in which this
validation ends does not delete the network and the images that were
created by the script.
For example if you want to create a new network on on the same device as
the "nova" one that is created by the validation, you get an error.
So I removed the deletes that were done before launching single commands
(since we're starting from a clean env there's no need to clean
anything) and I added a full clean on the cleanup function (called when
exiting).

Change-Id: I4342897dc88424a119a9668698668972caf03676