Persist ironic configuration changes
Quickstart sets the libvirt uri option for the ssh driver; However,
this is set in a way that is overwritten every time the undercloud is
updated. This commit changes the ironic configuration change to use a
supported medium instead, which is to use the hieradata_override
option. This then will allow us to update our undercloud without
issues.

Change-Id: Iff8bc6aaf629dd3027eac153578949a51d79d742
Closes-Bug: #1610205