Adding exit 1 after a failed deployment
After each deployment we always exit with success by using "|| true"
since we need, in case of failed deployments, to populate some log files
for further analysis. This commit adds a "1" exit status at the end of
the log operations, so the overall process will be interrupted as a user
may expect.

Change-Id: I88cf472e24fa7bd04cb99b6ff779de9f2aed8113