Add optional upstream ipxe installation
This commit adds the optional possibility to clone and install the
upstream ipxe repository, to use the latest version of ipxe and make
introspection on certain kind of hardware more reliable.
See this bug https://bugzilla.redhat.com/show_bug.cgi?id=1324422 for an
example in which using upstream ipxe makes things work.

Change-Id: I425f1404c4ab180182fc6c48ba61221e4cadda37