Updating default_floating_pool
tripleo-heat-templates changes the default_floating_pool from
nova to public in tripleo-heat-templates/puppet/services/nova-api.yaml.
So, while tempest creates the public network named nova, the controller
s expecting a public network named public.
Because of that, all the tests that relays on create a floating ip is
failing on tempest.
This fix the following tests:

- tempest.api.compute.floating_ips.test_floating_ips_actions
- tempest.api.compute.floating_ips.test_list_floating_ips
- tempest.api.compute.servers.test_server_rescue

Change-Id: Ib84444eca96209a094a94a9a80cd2611e1af44f6