Use rdopkg findpkg to map project names for DLRN build
With rdopkg 0.42 we get a new rdopkg findpkg command, that allows us
to map project names using rdoinfo. Now we don't need the
map-project-name script, and it is planned to be deprecated in the
future.

We need to add a step to fetch the rdoinfo repo, since previously
we relied on map-project-name doing it for us.

Change-Id: I70f38188694e3e5a26f0fa73072d1506326df1b0