Add generic undercloud setup role
All workflows not using the undercloud ready image
are required to setup the undercloud so that it
contains the elements required for installation and
deploy.

This role should include general undercloud set up
steps but can also include environment specific
steps that are only run under certain conditions.

Steps to get and modify overcloud images are in a
separate role - to be extracted from tripleo-quickstart.
Steps to get and enable repos are in a separate role.

Change-Id: Ibc5af913378ee6b713f94a775ea8ba8b6ffbfd0c