Modify OVB role to work with non-ready undercloud images
The OVB role was written at first to work with ready undercloud
images. This review changes the workflow to use a vanilla
guest image for the undercloud and set up the users, repos
and overcloud images after the OVB stack is deployed.

This streamlines the OVB workflow on different environments.

Change-Id: Ic8f40b5e38fab5aee3dc84d6fb1462b3f0f61ca7