Support a custom virthost key and update README
This commit adds the ability to pass a specific custom ssh private key
to quickstart so it can be able to connect to machines which are
customized inside the local .ssh/config file and to which you don't
connect with the default key of the user executing quickstart.
If the variable virthost_key is defined the initial ssh.config.ansible
is populated with this content:

      Host {{ virthost }}
          Hostname {{ virthost }}
          IdentityFile {{ virthost_key }}
          User root
          StrictHostKeyChecking no
          UserKnownHostsFile=/dev/null

So quickstart is able to connect to virthost, otherwise a connection
error would happen.
It also updated the README with all the explanations around this new
parameter and the fixes about the changes happened in the meantime
within the role.

Change-Id: Id9c79cf1c4aa3ff3b0c8484cebabe2b67faa691a