Remove `validation-warnings-fatal` with scale
HOTFIX: This flag, handed to tripelo when calling `overcloud deploy`
will kill deployment whenever a warning is encountered. Presently,
nodes that exist but are not used at the time of deployment trigger
a warning. This flag was introduced to OOOQ recently[1] and is
breaking scale periodic jobs in ci.centos. There is an upstream
review[2] that is in progress and will remove the unused node trigger
warning at which point this patch can be reverted.

[1] - https://review.openstack.org/#/c/404726/
[2] - https://review.openstack.org/#/c/414167/

Change-Id: I3e6bba38752c8689d4aedefdf8dd36cc90e71d09