undercloud-setup: Add a timeout to SSH for multinode vxlan setup
This commit addresses a problem of hung CI jobs during the execution of
toci_vxlan_networking_multinode.sh. Specifically, if incorrect
credentials are passed the job can hang indefinitely until the worker is
reaped. To do this a 'timeout' command is prepended to the script with a
default timeout of 60 seconds.

Change-Id: I2093ed1c7c4e33f96872153b66b7fc446b150fab