Add external repo and img roles to bm undercloud
This commit replaces the actual repo configuration and overcloud
images preparation inclusions with the already existing external
dedicated roles (repo-setup and fetch-images).
To configure this the two playbooks are modified and a new "packages"
subrole is created under the main baremetal-undercloud one.
With this in place it is possible to follow this sequence:

- Provide and prepare the undercloud machine (role baremetal-undercloud)
- Inventory the newly created machine (role tripleo-inventory)
- Setup repositories (role repo-setup)
- Install and configure packages on undercloud (baremetal-undercloud's
  subrole named "packages"):
  - tripleo package
  - additional packages (optional)
  - ntp configuration (replaces what was done before in the custom
    template undercloud-repos-conf.sh.j2)
- Install the undercloud

A new review to update the documentation will be done after this gets
merged.

Change-Id: Iba11c4d61a63bb05016177abd16403168d82331b