virtualbmc setup is not working in devmode.sh
Some workflows like devmode will not have facts
collected on the virthost e.g. devmode.sh. This
occurs when quickstart.sh is invoked multiple times.

Do not rely on facts to get the uuid of the
non_root_user, instead delegate to the virthost
and get the uuid and set a fact.

Closes-Bug: #1657248

Change-Id: I5d087983b1794c10c311bc44b162428c2cc7e476