[build-test-packages] combine all methods of providing changes
The current implementation of build-test-packages allows for
precisely one method of providing what changes to build.
Either passing a artg_change_list dict directly via variable,
reading GERRIT vars from the env, or reading ZUUL vars from
the env.

This patch makes it possible for multiple of these methods to
be used without stomping on eachother. It is needed to be able
to run the gate which specifically exercises this role, but may
be useful in other scenarios as well.

Change-Id: Ib2290411865f9bb1a2804af64e85541530ae2e97