Make validation HA tests atomic
This commit defines a var file for each release containing the list of
the HA validation tests that should be done. These new variables are now
also used as booleans to determine if a test should be executed or not.
With this new approach a user will be able to exclude (or include) tests
at runtime, like this:

./quickstart.sh \
  --retain-inventory \
  ...
  ...
  --release mitaka \
  --tags "overcloud-validate-ha" \
  --extra-vars test_ha_failed_actions=false \
  --extra-vars test_ha_master_slave=false \
  --extra-vars test_ha_keystone_constraint_removal=false \
  --extra-vars test_ha_instance=true \
  $VIRTHOST

In this case I will test just the instance deployment.
This means that without any extra var tests will be the default
depending on the release, otherwise specific tests will be made.
It also updates the README with all these new informations.

Change-Id: I113b8e929dddb9ff18b57faa905a754e2f6b5999