Collect and manage logs when validating HA
This commit manages all the stdout and stderr status of each HA
validation test.
In this way it will be possible to have all the logs related to the
tests in one place (on the undercloud) that will be collected by the
collect_logs role and so accessible from the external location where
the logs will be stored.

Change-Id: Ie803a920a17bd2f24ccd89f733781dd7e4d07fdd