Clone tripleo-quickstart dependant change to tripleo-quickstart-gate-repo
Currently the gating tripleo-quickstart-extras changes with
dependencies on tripleo-quickstart changes is  broken.
The reason being that the checkout task deletes the tripleo-quickstart
directory while it is in  use.
This review add a tripleo-quickstart-gate-repo to add the dependent
change to. After this role and quickstart run has completed,
the tripleo-quickstart-gate-repo directory can be renamed to
tripleo-quickstart before the deploy ci-script is run.

Change-Id: Ieaaf230155dcc2cdaafede63be193306eeb40499