Consider also newton for bulk introspection
This commit adds also newton release to be checked for the bulk
introspection.
Without this you won't be able to deploy because of this error:

Requested API version 1.6 is not supported by the server or the
requested operation is not supported by the requested version.
Supported version range is 1.1 to 1.22

Things get worse because this error does not block the execution of
quickstart, because the related command is executed in $() so it does
not obey to the set -eux on the top of the script, making the process
going forward and failing the overcloud deployment with the "No valid
host found" infamous error.

Change-Id: I5740b908bac421356bbd7b6f3fa8104f21b5cf25