Add scenario 002-005 config files for multinode scenario jobs
This commit adds config files for the remaining scenario jobs. Although
these are nearly identical, they should in theory change which logs are
collected to reflect the different services
deployed by each scenario.

It also fixes a comment at the beginning explaining why introspection
is disabled. It also removes an extraneous compute_0 host from the list
of overcloud nodes in the config files.

Change-Id: I581cb3e99e9f38f083ec9f55c9e5d9a1643a0e17