Fix introspection with retries edge case
https://review.openstack.org/#/c/403677/ finally landed and started
running in my jobs and builds started failing spuriously.

Long story short my manual test case was not representative of all actual
use cases.

For large clouds by the time we get to the timeout while loop at least
one node has always already finished introspection and the script works
as advertised. For smaller clouds (like my 3 node CI clouds which don't
really need this script very badly) this is not the case and the script
will fail.

To solve this problem I now explicitly wait for at least one node to be
on and in manageable state before going into the timeout and remove the
on != node count check entierly.

tl;dr test code that needs to run on 50 node clouds and 2 node clouds on
both. Yes I did that this time

Change-Id: Ie0f28567ec5713241d2ca0ed0ed2598d6d769b44