Add timestamps to quickstart logs
Prepend a timestamp with awk to our deployment and test commands. This
allows html linking from the log servers and also better correlation
with the system logs.

We are not using the "ts" command from the moreutils package to avoid
extra dependencies. The drawback is the lack of sub-second timinig.

It is possible to disable the timestamping by setting the
timestamper_cmd to an empty string.

Also add a README file for the extras-common role.

Closes-Bug: #1673046
Change-Id: I8a2a1e4fde2ba805d3361e2e7a2337e873133ef0