Do not disable other repos when updating gate packages
If we have a dependency that is not installed at the point we run
this, and it is in another repo, we will fail to find it and fail
the script.

DLRN already has a mechanism to guarantee a later built package has
a higher version than earlier packages, and we have priority set to
1 for this repo as well. As such we should get the packages we want
without disabling the other repos.

Change-Id: Id74954a8bab049f39af9f73bfcd8137b0d6f8139