Use upload command to upload container images
This switches to using the "openstack overcloud container image upload"
command instead of a custom script. They both use the same
tripleo-common library code.

This change removes the retry loop on failure. We believe these
failures were caused by the old docker-registry service, which has
since been replaced with the registry that comes with the
docker-distribution service, so I think it is worth trying without a
retry loop for now.

Change-Id: I618bc6cae400cbd441a2e72c02c20830e493c837
Depends-On: Id505b57698b6a15c62494764b2d9eca1933e63f1