Move OVB settings to tripleo-quickstart-extras
The OVB settings were kept in an internal repo.
Since settings for public clouds can be added to
the extras repo, where open settings can be used,
access to private repos should not be needed.

Note that the user will still have to source a rc
file to access resources on a tenant.

Change-Id: Id625c2a60811fafb353519ad9bf976fc7bdea826