Clone projects if they aren't cloned by ZUUL
When we build projects in ZUUL_CHANGES we usually have them
cloned by infra with zuul-cloner in /opt/stack/new with all
required patches, so we just copy them and build. But sometimes
we need to build dependencies in projects that are not in PROJECTS
and not cloned, like heat, glance, etc. This commit supports
cloning and patching of projects that are not cloned by ZUUL
initially.
It was partially implemented in tripleo.sh so we close the gap.

Close-Bug: #1675380
Change-Id: Icf1af819a07d47eaad185b02795bc68de9cc8e32