Address multiple gate issues
Add filter for zuul_deps

Add the ability to filter which changes to return. As an example, if
you are running on a stable/neuton branch, you might not be interested
in changes from master.

---

Only parse change for the ZUUL_BRANCH we care about

We should only build packages for the ZUUL_BRANCH a job is on. As an
example, if the patch is for stable/newton, I don't believe we want to
build DLRN packages for master.

---

Reverse list of zuul_changes

In the following case, ZUUL_CHANGES contains:

  openstack-infra/tripleo-ci:master:refs/changes/23/449023/2^\
    openstack-infra/tripleo-ci:master:refs/changes/41/448041/2^\
    openstack/tripleo-heat-templates:master:refs/changes/08/444308/1

We actually want to drop refs/changes/23/449023/2 because appears
refs/changes/41/448041/2 appears later in the change queue.  Meaning,
448041/2 contains 449023/2 already. Otherwise, we'll never build
448041, not testing the proper checkout.

Change-Id: Ic7e20df98d690d268081d395f93a8a098e6b3d95
Signed-off-by: Paul Belanger <pabelanger@redhat.com>