undercloud/heat: Set workers correctly for httpd
With a recent commit [1], the heat services are now deployed over httpd.
This commit reflects this change when we try to set the workers for
heat, since the previous parameters don't work anymore.

Note that this change applies to master (pike), so previous releases
should still use the previous worker setting.

[1] I77b587cd636594273931586538a11b8936588a76

Change-Id: Iceb9073d542af20273326a7791cd20840d7f5dbd