Fix stackviz data generation
After I177b32622500dab54896fda740080d07f5f65261 we started to use
testrepository from system, which is version 0.0.18 and generate subunit
v1 output, and stackviz doesn't recognize it. So it's required to
transform version 1 to 2 with subunit-1to2 command.
I've tested manually, and the subunit-1to2 works either if you're using
version 0.0.18 or 0.0.20.

Change-Id: Ie7c3f3e734fabdd116cb2421e85c0595e3c56a87