Use web repo with installed packages for image build
When building images we need to supply repo with built packages
to DIB so it will use when installing packages on the image.
Because DIB is in a chrooted environment we have to use repo in
network. This patch runs a webserver on the directory of the
repository to make it accessible over HTTP on port 8766.

Change-Id: I7b21d3ff1ec7cecab322e599cb929259f5a7256e