Allow depending on changes we dont actually build
It is common practice in TripleO to depend on a documentation
change when changing some aspect of CI or dev tooling. However, we
can not actually build a package from these changes, so the job
will fail.

This change makes sure we only try to build the package if it maps
to some package in rdoinfo.  A debug message will be printed if no
package is found.

Change-Id: I076d1ed89bc8707630841cbcb2683295e935a944
Closes-Bug: 1679650
Co-Authored-By: Alex Schultz <aschultz@redhat.com>