Increase default novajoin read timeout to 20 seconds
If the timeout is too short then retrieving metadata from within
an instance can fail. Given that several calls to IPA may be made
during the request 5 seconds is too short to guarantee that
the IPA work will complete (which will include doing a GSSAPI
request and TLS connection).

Failure to retrieve the metadata will result in a failed IPA
enrollment on the node which will result in an overcloud deploy
failure.

Change-Id: I684118631cfa623e3274ffaa22936f34389d753a