Rename misleading undercloud/overcloud hieradata overrides
There was a misconception and thus it ended up in a misleading name for
the hieradata overrides set up by quickstart. Thus, the hieradata
overrides that were there before ended up being called the "overcloud"
hieradata, even though it's not used in the overcloud at all. In
reality, it was being used by the classic undercloud install and is used
to configure extra bits from the undercloud OpenStack deployment.

Thus I changed the names to reflect the actual meaning of these files.

Change-Id: I6504f03706c81becea5c38b03ef619a5613c544c