Don't overwrite tht_dir for upgrades to master
Overwriting this broke mixed version upgrade [1], which tried to copy
t-h-t to `/home/stack//usr/share/openstack-tripleo-heat-templates`:

https://github.com/openstack/tripleo-quickstart-extras/blob/b1a1f010db17a4abf9281748ee36bcbe0eee6c88/roles/overcloud-upgrade/tasks/pre-undercloud.yml#L39

Setting tht_dir like this is suspicious given how tht_dir is used in
rest of the code, and this should probably be fully removed eventually,
but i'm keeping it at least for Ocata as i suspect that there might be
some jobs relying on the semi-broken behavior (which probably aren't
running a mixed version upgrade, otherwise they'd fail).

In case we are using a composable scenario which didn't exist
previously, we copy in the composable scenario from the RPM location, so
that OOOQ can find the expected environment file without needing to
backport composable scenario env files to older releases.

Change-Id: Ib3fdaa87c211545f835afb1c95e4d13debba2827