Support arbitrary nat network order
Previously the vbmc would use the address of the first network
defined in order to connect. By changing this to search for the
first network with an address and fall back to using the default
libvirt network address, networks can be re-ordered and re-named
as desired.

Default libvirt network support is added in
https://review.openstack.org/#/c/406421/

Change-Id: Id4a298f91c75bcd0843648939c0a8da02b9f6a6c