Add stack deploy status output for failed deploy debugging
If an OVB stack does not deploy successfully, the os_stack module
fails out without returning any reason for failure.
This review ignores the error returned by os_stack and
adds a debug step that prints out the stack status so the
user knows why the deploy failed.

This review deletes the os_stack module stored in
ovb-manage-stack role as it is now included with
Ansible.

Change-Id: I7f0d4ef68d06cddf18101da20877cf49e8415069