Explicity use bash for ovb-manage-stack shell block
"set -o pipefail" is a bashism, and it is not within the POSIX
standard. It is possible that the system's default shell (/bin/sh) is
not bash, which can result in errors.

This change only corrects this in ovb-manage-stack which is most likely
to directly run on non-CentOS/Fedora/RHEL systems during the devmode OVB
execution.

Change-Id: I7a98001f125ef903766543102cd0506a243606f3