overcloud validate: Increase ping count to 3
When the ping test runs (in the undercloud) to ping to the
floating ip of the VM, ping count of 1 is not sufficient
and it could fail for some SDN controllers. In the case of
OVN (wherein it has native L3 support and unlike neutron
reference implementation it doesn't create router namespaces),
when the VM sends the icmp (ping) response, the OVN router
pipeline generates an ARP packet to learn the mac address
of the next hop (if it hasn't already learnt yet, which is always
the case for the fist packet) causing the first ping reply packet
to be lost. Also there is no harm in increasing the count to 3.

Change-Id: I4879f2f3cecced8b96e4ee230e4de7b210a8f7f8