Adding real time bug verification
This patch adds a script that will check all the bugs in the
yaml skip list file which has a bugzilla or launchpad attached
and verify if the bug still open or closed. If closed, it will
remove from the skip list.
This also can generate a report showing what is still open and
what was already closed, as well as generate the skiplist in
tempest/testr format.

Change-Id: Iadd508184c350d074600316ad15f5274df87f01b