Use prepare command for overcloud_containers.yaml
Most deployments will not be able to use the packaged
overcloud_containers.yaml unmodified, so the prepare command will
allow local modified files to be created.

Once nothing in CI consumes the packaged overcloud_containers.yaml, it
will be deleted.

This change also used the --env-file option to generate
container-default-parameters.yaml instead of the current templated
approach. There is about to be some disruption to how the namespace is
represented in the image parameters, so using the prepare command will
help mitigate that.

Change-Id: I6a5796428c224584359b2a434d7fc80878d5f038
Partial-Bug: #1696598