Allow removing of options from tempest conf
This patch gives the option to pass parameters to tempest to
remove certain keys while generating tempest.conf by passing
it as tempest_conf_removal.

The tempest_conf_removal is applied in the final stage and
as such overrides even if values were prevously set.

dvr is default value in tempest_conf_removal - neutron extension.
python-tempestconf gets all available neutron extensions, even if
they are not enabled in a deployment. dvr is not generally configured
so to avoid failed tests, it's removed from tempest.conf

Change-Id: I708866aa3218b2e3c25e9b47672237bc1838d760