Pacemaker args must be included before containers
Today if we want a containerized highly available deployment with
quickstart we need to override the command line parameters inside the
"container_args" variable.
By logic, if one needs HA he will include also "pacemaker: true" in his
config, and this will populate also "pacemaker_args" variable. The
defaults of the overcloud-deploy role are set so that on the deploy
command line pacemaker_args comes after container_args. This breaks
things since the included pacemaker yaml overrides the most of the
OS::TripleO::Services inclusions. Putting it after would avoid this.
Another solution is to put "pacemaker: false" so that no inclusion is
made, but it's somehow illogical.

Change-Id: I9d8781621bdc0b776165a790ea493ab5cfcbd946