Make env variables for multinode into ansible vars
This will allow us to configure what is passed to the deployed_server
script dynamically.

The defaults match what we currently use for the single node deployed
server jobs. We can consider changing the default to something less
CI specific when we have the ability to run the deployed server jobs
on libvirt.

Change-Id: I55f705a390ae5b927a9fb0f8d8ec153ea61cc3c9