Fix conditional on repo_built
As per log in [1] and some local jobs, a register in a
block as repo_built in [2] is always registered even if
the task is skipped. in that case repo_built will have
value:

    "repo_built": {
        "changed": false,
        "skip_reason": "Conditional check failed",
        "skipped": true
    }

so any conditional checking "repo_built is defined" will
pass even if the task registering the variable is skipped.

This patch adds a condition to check if the repo_built.skipped
is not defined.

[1] https://review.rdoproject.org/jenkins/job/rdoinfo-tripleo-ocata-testing-centos-7-multinode-1ctlr-featureset005-nv/19/consoleFull
[2] https://github.com/openstack/tripleo-quickstart-extras/blob/master/roles/build-test-packages/tasks/dlrn-build.yml#L88

Change-Id: Ib0b313cd1244de6a3916cd691e7b824b4f952f80