Don't pass puppet-pacemaker.yaml to containerized HA envs
This shouldn't change functionality, as container_args would override
values from this env file when both containerized_overcloud and
enable_pacemaker are set to true. However, we wouldn't be recommending
users to deploy with both `-e puppet-pacemaker.yaml -e docker-ha.yaml`
at the same time, so we shouldn't be passing it like that in OOOQ
either.

The change is from a patch by Wes [1], instead of which we merged a
different patch in the end, but i believe we still need at least this
part from it.

[1] https://review.openstack.org/491146

Change-Id: I9e0858b5cee8d14e69f218cd9957079787b92db8
Co-Authored-By: Wesley Hayutin <weshayutin@gmail.com>
Related-Bug: #1712070