tempest/exclude: remove TestEncryptedCinderVolumes
TestEncryptedCinderVolumes test shouldn't be ignored and the test should
work correctly. It was ignored as a workaround because test was *always*
running but this was not the right solution. If the test is running
while we don't want to, we need to find the root cause and fix it.

In the meantime, let's enable this test that we need on scenario002
where Cinder and Barbican are enabled.

Change-Id: I0ff91da4420f81cac6177b57f173b8305617aa4d