Sudo fix overcloud-custom-tht-script.sh.j2
Using overcloud-custom-tht-script.sh.j2 to use a git clone of tht
currently fails because the user lacks permission to modify
/usr/share.

Also, having the git clone directly into /usr/share is not very useful
for those using quickstart for tht development.

This change does a git clone of tht to the working directory and symlinks this
to /usr/share/

It also makes $HOME readable so that mistral workflows have permission
to read the local tripleo.

Change-Id: I08f259664c7810f3a1b1373285bfec1227144da9