Allow composable scenarios without deployed server environment
Whenever we deploy scenario with deploy_composable_scenario, OOOQ
assumed we're deploying on deployed server (which doesn't have to be
the case) and it included
deployed-server-bootstrap-environment-centos.yaml.

The deployed server environment should be part of
multinode_args (where the other env file for deployed server already
is) rather than scenario_args.

Change-Id: Ibeb0b964552c2ad3468c3f22bb7e3a4e7543849a
Closes-Bug: #1713958