Use different variables for deploy and upgrade scenarios
Introduce an upgrade_composable_scenario variable to hold the path to
the scenario file used in the upgrade command to differentiate with the
composable_scenario variable for the path of the scenario file used in
deploy command. Remove the deploy_composable_scenario variable that was
previously holding a boolean value, which can be inferred from the
content of the composable_scenario variable.

Reorder the arguments passed to the deploy command so that the default
from the docker.yaml environment doesn't overwrite the setting coming
from the scenario.

Co-Authored-By: Jiří Stránský <jistr@redhat.com>
Depends-On: Ie08bbe08530bd48a0ca58667f0704f360e0a4dd7
Change-Id: Iece70e22cd40cb119f2f9eb3c51be7bd4631fade
Related-Bug: #1714905