Increase session IPMI timeout in VBMC
The default IPMI session timeout is 1s. This is rather
low if you use oooq to deploy a big overcloud on a biggish
server. Let's increase this timeout by a large amount so
to give plenty of time to vbmc to operate.

With this change the ironic errors I observe from time to
time on my deployments are gone.

Change-Id: I37afb2e7d9e57405b1c34b413f8152bed97a344a