Remove BASE var from vxlan_networking.sh
The /opt/stack variable was being used in a single
place and that use can be easily substituted
by the tripleo_root variable, which is the place
were the repositories are being downloaded. If
for some reason tripleo_root directory does not
end up in /new then an error appears as it won't
find /opt/stack/new/devstack/functions-common.

Change-Id: Ibff5e51c0b7a06b9915e222712738d6ca93a366f