Make sure ansible_distribution is defined
If the ansible_distribution fact is not gathered, then the
overcloud-deploy role will just fail while trying to
evaluate it.

Instead we just make sure it is defined before trying to
evaluate it. This will have the consequence of any deploy
which cannot check the ansible_distribution not getting the
libvirt_args set unless by explicit config. However, that
seems better than failing outright.

Closes-Bug: #1721790

Change-Id: I08c689b949372bae9ffd7eb4ce9576c5b37d138d