Split general OVB setup tasks from the the stack create
This is the first step in combining the upstream playbook used
for OVB jobs with this one. This should be a no-op change in terms
of what gets run, however the part that is split out is common to
the upstream jobs which create the stack via the testenv broker.

We can then iterate on adding in the testenv broker steps (guarded
by some option) in follow-up patches.

Change-Id: I7cad64e83ed0fbf791b26bdf0e381599d177de7b