Translate tripleo-ci logic of when to build images to ansible
In order to combine the tripleo-ci OVB playbook with what we have
here, we need to implement the logic around when to build images[1].

We default to fetching pre-built images. However, in the case we are testing
a change to a project that can affect the image build or the case we are
running a periodic job, we want to build the images.

[1] https://github.com/openstack-infra/tripleo-ci/blob/master/scripts/to_build

Change-Id: I72169d4c6ae891397916f831e56eb883d047c89b