Add steps needed when using OVB with a testenv broker
This continues the work to merge the OVB playbook in tripleo-ci with
the OVB playbook here. This should again be a no-op change, with the
only new behavior turned on by setting use_testenv_broker=true.

This change completes the differences between the two playbooks before
the undercloud is installed. There are also slight differences in what
we do after the undercloud is installed in the
baremetal-quickstart-extras playbook. Those will be addressed in the
final patch of this series.

Change-Id: Id24b8c4866f46dcb37a98b682f54101a341f9adb