Add the option to run the container-check script
This review adds the functionality to install the
container-check package and run the script
to update the containers and include the latest rpms.
The containers should be updated when there is a change
to be built and tested or when the latest rpms are required.

By default, containers will not be updated. Feature sets or
environment files will be used to enable the container update.

Change-Id: I7fc3a83c95a76fe5005eeb084ad044fee9084bdb