Provide a quick workaround for 1727454
Tripleo-common provides an overcloudrc and an overcloudrc.v3 by default.
Problem is that Keystone v2 API was removed in Queens, so we can't rely
on overcloudrc file anymore.

Until we fix that in tripleo-common, let's source the overcloudrc.v3
after Pike, so we can promote to Queens.

This is technical debt that will be cleaned asap.

Change-Id: If795dfe6953dd7a68bbf075f476f8687c06420a1
Related-Bug: #1727454