Install openstack-selinux with openvswitch
Since openvswitch-2.8, some selinux policies installed
with openstack-selinux are required for openvswitch
services to start properly.

Funtion ovs_vxlan_bridge is used in multinode environments
to instal ovs in subnodes, so this patch is installing
openstack-selinux when installing openvswitch. Note that
selinux is not mandatory in ovs for non-openstack environments
so it can't be added as dependency in RPM.

Change-Id: I47832c2cd84247f0ee0e02208e95c051087c494a